changing filter price product admin adding products

+++ src/Entity/Product.php
@@ -22,7 +22,7 @@ class Product
     #[ORM\Column(type: 'string', length: 120)]
     private $img;
 
-    #[ORM\Column(type: 'integer')]
+    #[ORM\Column(type: 'float')]
     private $price;
 
     #[ORM\ManyToOne(targetEntity: Category::class, inversedBy: 'product')]
@@ -38,7 +38,7 @@ class Product
     #[ORM\Column(type: 'string', length: 255)]
     private $statutes;
 
-    #[ORM\Column(type: 'string', length: 255)]
+    #[ORM\Column(type: 'string', length: 4000)]
     private $ingredients;
 
     #[ORM\Column(type: 'integer')]
@@ -85,12 +85,12 @@ class Product
         return $this;
     }
 
-    public function getPrice(): ?int
+    public function getPrice(): ?float
     {
         return $this->price;
     }
 
-    public function setPrice(int $price): self
+    public function setPrice(float $price): self
     {
         $this->price = $price;
